requirements.txt
albumentations==1.1.0
numpy==1.17.4
opencv_python==4.4.0.46
Pillow==9.1.0
torch==1.10.2
torchvision==0.11.3